pyproject.toml
[project]
name = "upbit-auto"
version = "0.1.0"
description = "Automated crypto trading and backtesting for Upbit (CLI: ua)"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "upbit-auto" }]
license = {text = "Proprietary"}
keywords = ["upbit", "crypto", "trading", "backtest", "cli"]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: Other/Proprietary License",
  "Environment :: Console",
  "Topic :: Office/Business :: Financial :: Investment",
]

dependencies = [
  "typer>=0.12.3",
  "httpx>=0.27.0",
  "websockets>=12.0",
  "pandas>=2.2.2",
  "numpy>=1.26.0",
  "backtesting>=0.3.3",
  "pydantic>=2.6.4",
  "structlog>=24.1.0",
  "orjson>=3.10.7",
  "python-dotenv>=1.0.1",
  "PyJWT>=2.9.0",
]

[project.scripts]
ua = "ua.__main__:app"

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
